move login code from auth to login.js and removed login code we are no longer using
