+++ views/login.hbs
@@ -9,7 +9,7 @@
                     {{message}}
             </div>
             {{/if}}
-        <form class = "form" action="auth/login" method="POST">
+        <form class = "form" action="/login" method="POST">
             <div class="form-group">
                 <label for="email">Email address</label>
                 <input name="email" type="email" class="form-control" id="email" placeholder="Enter email">
@@ -27,5 +27,3 @@
 </div>
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script src="/js/shared.js"></script>
-    <script src="/js/login.js"></script>
